Began to separate main file to make it easier to manage
